+++ src/lib/components/asset/SupportedAssetTitle.tsx
@@ -19,14 +19,7 @@ export const SupportedAssetTitle = ({
     !totalSupportedAssetsValue || totalSupportedAssetsValue.eq(0);
 
   return (
-    <Flex
-      w="full"
-      bg="gray.900"
-      py={2}
-      px={4}
-      borderRadius="8px 8px 0px 0px"
-      justifyContent="space-between"
-    >
+    <Flex w="full" py={1} gap={2} align="center" justifyContent="space-between">
       <TableTitle
         title="Supported Assets"
         count={supportedAssets.length}
@@ -37,7 +30,7 @@ export const SupportedAssetTitle = ({
         <Heading
           mt={1}
           as="h6"
-          variant="h6"
+          variant="h7"
           color={isZeroValue ? "text.dark" : "text.main"}
         >
           {totalSupportedAssetsValue